backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
scikit-fuzzy
scipy
numpy
httpx
python-dotenv
networkx
packaging
